requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
requests==2.18.1
rosette-api==1.7.0
urllib3==1.21.1
